requirements.txt
groq==0.4.2
httpx==0.27.0
python-dotenv==1.0.1
streamlit==1.32.2
streamlit-authenticator==0.3.1
pyyaml==6.0.1
requests==2.31.0
mcp-tools==0.1.0
